Fix plugin install scanner validation
