+++ openclaw.plugin.json
@@ -2,7 +2,7 @@
   "id": "rocketchat",
   "name": "Rocket.Chat",
   "description": "Rocket.Chat channel plugin",
-  "version": "0.3.3",
+  "version": "0.3.4",
   "channels": [
     "rocketchat"
   ],
