+++ tsup.config.ts
@@ -8,5 +8,5 @@ export default defineConfig({
   clean: true,
   outDir: "dist",
   target: "es2022",
-  external: ["openclaw", "openclaw/plugin-sdk", "ws"],
+  external: [/^openclaw(?:\/.*)?$/, "ws"],
 });
